Add song playback test.
